op-node: Simplify Receipts Fetching API
The previous API would return a receipts fetcher that the caller
was responsible for iterating through and then calling `Result` on.
This moves this inside the `Fetch` function. This simplifies usage
of the API and does not result in performance regression because
the receipts hash check requires all receipts to be fetched before