Cache receipts fetcher rather than receipts
This modifies the refactor to be equivalent to the previous code.
Caching the receipts fetcher is actually critical for performance
because the fetcher caches intermediate results. The fetcher typically
makes lots of calls to fetch all the receipts so if all work is thrown
out on a single failure it will be very hard for the node to make progress.